Refactor matchers file layout
